+++ authorization-service/src/utils/auth.ts
@@ -1,11 +1,17 @@
-export class Auth {
-    static decodeToken(token: string) {
-        return Buffer.from(token, 'base64').toString('ascii');
+export const generateAuthPolicy = (
+    principalId: string,
+    resource: string,
+    effect: string = 'Allow'
+) => ({
+    principalId,
+    policyDocument: {
+        Version: '2012-10-17',
+        Statement: [
+            {
+                Action: 'execute-api:Invoke',
+                Effect: effect,
+                Resource: resource
+            }
+        ]
     }
-
-    static isAuthorized(token: string) {
-        return (
-            token === `${process.env.USER_NAME}:${process.env.USER_PASSWORD}`
-        );
-    }
-}
+});
