--- authorization-service/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './policy';
-export * from './auth';
