+++ authorization-service/tsconfig.paths.json
@@ -3,6 +3,7 @@
         "baseUrl": ".",
         "paths": {
             "@functions/*": ["src/functions/*"],
+            "@services/*": ["src/services/*"],
             "@utils/*": ["src/utils/*"],
             "@libs/*": ["src/libs/*"]
         }
